+++ firehosepool/connect.go
@@ -99,5 +99,5 @@ func (srv *Server) clientsReset() {
 			srv.clients = append(srv.clients, NewClient(srv))
 		}
 	}
-	metricWorkers.WithLabelValues(srv.cfg.StreamName).Set(float64(len(srv.clients)))
+	metricWorkers.WithLabelValues(srv.cfg.Label).Set(float64(len(srv.clients)))
 }